Fixing tests

+++ app/models/project_media.rb
@@ -244,32 +244,36 @@ class ProjectMedia < ApplicationRecord
   end
 
   def replace_by(new_pm)
+    return if self.id == new_pm.id
     if self.team_id != new_pm.team_id
       raise I18n.t(:replace_by_media_in_the_same_team)
     elsif self.media.media_type != 'blank'
       raise I18n.t(:replace_blank_media_only)
     else
-      # Save the new item
-      new_pm.updated_at = Time.now
-      new_pm.skip_check_ability = true
-      new_pm.channel = { main: CheckChannels::ChannelCodes::FETCH }
-      new_pm.save(validate: false) # To skip channel validation
+      ProjectMedia.transaction do
+        # Save the new item
+        new_pm.updated_at = Time.now
+        new_pm.skip_check_ability = true
+        new_pm.channel = { main: CheckChannels::ChannelCodes::FETCH }
+        new_pm.save(validate: false) # To skip channel validation
 
-      # Point the claim and consequently the fact-check
-      new_pm = ProjectMedia.find(new_pm.id)
-      new_pm.claim_description = self.claim_description
-      new_pm.save!
+        # Point the claim and consequently the fact-check
+        new_pm = ProjectMedia.find(new_pm.id)
+        new_pm.claim_description = self.claim_description
+        new_pm.save!
 
-      # All annotations from the old item should point to the new item
-      # Remove any status and report from the new item
-      Annotation.where(annotation_type: ['verification_status', 'report_design'], annotated_type: 'ProjectMedia', annotated_id: new_pm.id).delete_all
-      Annotation.where(annotated_type: 'ProjectMedia', annotated_id: self.id).update_all(annotated_id: new_pm.id)
+        # All annotations from the old item should point to the new item
+        # Remove any status and report from the new item
+        Annotation.where(annotation_type: ['verification_status', 'report_design'], annotated_type: 'ProjectMedia', annotated_id: new_pm.id).delete_all
+        Annotation.where(annotated_type: 'ProjectMedia', annotated_id: self.id).update_all(annotated_id: new_pm.id)
 
-      # All versions from the old item should point to the new item
-      Version.from_partition(self.team_id).where(associated_type: 'ProjectMedia', associated_id: self.id).update_all(annotated_id: new_pm.id)
+        # All versions from the old item should point to the new item
+        Version.from_partition(self.team_id).where(associated_type: 'ProjectMedia', associated_id: self.id).update_all(associated_id: new_pm.id)
+        Version.from_partition(self.team_id).where(item_type: 'ProjectMedia', item_id: self.id).update_all(item_id: new_pm.id)
 
-      # All relationships from the old item should point to the new item
-      Relationship.where(source_id: self.id).update_all(source_id: new_pm.id)
+        # All relationships from the old item should point to the new item
+        Relationship.where(source_id: self.id).update_all(source_id: new_pm.id)
+      end
 
       # Clear cached fields
       new_pm.clear_cached_fields
@@ -278,20 +282,22 @@ class ProjectMedia < ApplicationRecord
       Rails.cache.write("check_cached_field:ProjectMedia:#{new_pm.id}:creator_name", 'Import')
 
       # Apply other stuff in background
-      self.delay_for(5.seconds).apply_replace_by(self.id, new_pm.id)
+      self.class.delay_for(5.seconds).apply_replace_by(self.id, new_pm.id)
     end
   end
 
-  def apply_replace_by(old_pm_id, new_pm_id)
+  def self.apply_replace_by(old_pm_id, new_pm_id)
     old_pm = ProjectMedia.find(old_pm_id)
     new_pm = ProjectMedia.find(new_pm_id)
 
     # Send the old item to the trash
+    Dynamic.create!(annotation_type: 'verification_status', annotated: old_pm, set_fields: { verification_status_status: new_pm.status }.to_json)
+    old_pm.clear_cached_fields
     old_pm.archived = CheckArchivedFlags::FlagCodes::TRASHED
     old_pm.skip_check_ability = true
     old_pm.save!
 
-    # Re-index both old and new items
+    # TODO: Re-index both old and new items in ElasticSearch
 
     # Send a published report if any
     ::Bot::Smooch.send_report_from_parent_to_child(new_pm.id, new_pm.id)
